pyproject.toml
[project]
name = "qa-repo-eval-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1",
    "requests>=2.31",
    "openai>=1.93.0",
    "gitingest>=0.1.5",
]
